Update 150219.1
Added Killfeed and new assets.
+As always fixed some stuff...

+++ Assets/Scripts/Game/GameSettings.cs
@@ -3,9 +3,11 @@
 [CreateAssetMenu(fileName = "New Game Setting", menuName = "Game/Setting")]
 public class GameSettings : ScriptableObject
 {
-    public int pointsAtStart = 0;
-    public int pointsToComplete = 50;
+    public short pointsAtStart = 0;
+    public short pointsToComplete = 50;
     public int pointsMultiplier = 1;
 
     public bool teamAssignEnabled = true;
+
+
 }
